+++ hutool-crypto/src/test/java/cn/hutool/crypto/test/SM2Test.java
@@ -6,6 +6,7 @@ import org.junit.Assert;
 import org.junit.Test;
 
 import cn.hutool.core.util.CharsetUtil;
+import cn.hutool.core.util.HexUtil;
 import cn.hutool.core.util.StrUtil;
 import cn.hutool.crypto.SecureUtil;
 import cn.hutool.crypto.SmUtil;
@@ -28,7 +29,7 @@ public class SM2Test {
 	}
 
 	@Test
-	public void rsaCustomKeyTest() {
+	public void sm2CustomKeyTest() {
 		KeyPair pair = SecureUtil.generateKeyPair("SM2");
 		byte[] privateKey = pair.getPrivate().getEncoded();
 		byte[] publicKey = pair.getPublic().getEncoded();
@@ -60,7 +61,7 @@ public class SM2Test {
 	@Test
 	public void sm2BcdTest() {
 		String text = "我是一段测试aaaa";
-		
+
 		final SM2 sm2 = new SM2();
 
 		// 公钥加密，私钥解密
@@ -68,20 +69,47 @@ public class SM2Test {
 		String decryptStr = StrUtil.utf8Str(sm2.decryptFromBcd(encryptStr, KeyType.PrivateKey));
 		Assert.assertEquals(text, decryptStr);
 	}
-	
+
 	@Test
 	public void sm2Base64Test() {
 		String textBase = "我是一段特别长的测试";
 		String text = "";
-		for(int i = 0; i < 100; i++) {
+		for (int i = 0; i < 100; i++) {
 			text += textBase;
 		}
-		
+
 		final SM2 sm2 = new SM2();
-		
+
 		// 公钥加密，私钥解密
 		String encryptStr = sm2.encryptBase64(text, KeyType.PublicKey);
 		String decryptStr = StrUtil.utf8Str(sm2.decryptFromBase64(encryptStr, KeyType.PrivateKey));
 		Assert.assertEquals(text, decryptStr);
+	}
+
+	@Test
+	public void sm2SignAndVerifyTest() {
+		String content = "我是Hanley.";
+
+		final SM2 sm2 = new SM2();
+
+		byte[] sign = sm2.sign(content.getBytes());
+		boolean verify = sm2.verify(content.getBytes(), sign);
+		Assert.assertTrue(verify);
+	}
+
+	@Test
+	public void sm2SignAndVerifyUseKeyTest() {
+		String content = "我是Hanley.";
+
+		KeyPair pair = SecureUtil.generateKeyPair("SM2");
+
+		final SM2 sm2 = new SM2(//
+				HexUtil.encodeHexStr(pair.getPrivate().getEncoded()), //
+				HexUtil.encodeHexStr(pair.getPublic().getEncoded())//
+		);
+
+		byte[] sign = sm2.sign(content.getBytes());
+		boolean verify = sm2.verify(content.getBytes(), sign);
+		Assert.assertTrue(verify);
 	}
 }