+++ hutool-extra/src/main/java/cn/hutool/extra/template/engine/freemarker/SimpleStringTemplateLoader.java
@@ -1,0 +1,38 @@
+package cn.hutool.extra.template.engine.freemarker;
+
+import java.io.IOException;
+import java.io.Reader;
+import java.io.StringReader;
+
+import freemarker.cache.TemplateLoader;
+
+/**
+ * {@link TemplateLoader} 字符串实现形式<br>
+ * 用于直接获取字符串模板
+ * 
+ * @author looly
+ * @since 4.3.3
+ */
+public class SimpleStringTemplateLoader implements TemplateLoader {
+
+	@Override
+	public Object findTemplateSource(String name) throws IOException {
+		return name;
+	}
+
+	@Override
+	public long getLastModified(Object templateSource) {
+		return System.currentTimeMillis();
+	}
+
+	@Override
+	public Reader getReader(Object templateSource, String encoding) throws IOException {
+		return new StringReader((String) templateSource);
+	}
+
+	@Override
+	public void closeTemplateSource(Object templateSource) throws IOException {
+		// ignore
+	}
+
+}